+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Feature - Add `:defaults_mode` configuration - that determines how certain default configuration options are resolved in the SDK.
+
 3.124.0 (2021-11-30)
 ------------------